almost finished fixing web service
the code I’m writing is crap!

+++ assets/linker/js/wService/wService.js
@@ -20,7 +20,7 @@ wService.controller('wServiceCtrl', ['$scope','$route', '$routeParams','$locatio
         $scope.setObjectData=function(wS){
             $scope.name=wS.name;
             $scope.bigImage='/images/'+wS.image+'_big.png';
-            $scope.smallImage='/images/'+wS.image+'_big.png';
+            $scope.smallImage='/images/'+wS.image+'_small.png';
             $scope.id=wS.id;
             
            // console.log('init',wS);
@@ -31,37 +31,69 @@ wService.controller('wServiceCtrl', ['$scope','$route', '$routeParams','$locatio
 
             var wService=WService.salva({
               project:projectid, 
-              name: "New Object"}, 
+              name: "New Web Service"}, 
 
-              function(){
-                console.log(wService);
-                $scope.$parent.wServices.push(wService);
-                $scope.$parent.selectCobject(wService);
-             //  console.log($scope.$parent.selectedWService);
+              function(w){
+
+
+                var newWservice={name:w.name, 
+                                bigImage:w.bigImage, 
+                                id:w.id, 
+                                image:w.image, 
+                                positionX: w.positionX,
+                                positionY: w.positionY,
+ }
+
+
+                console.log(newWservice);
+
+                //add newWservice to entities
+                $scope.$parent.entities.push(newWservice);
+                console.log("entities",$scope.$parent.entities);
+
+                //add newWservice to wservices
+                $scope.$parent.wServices.push(newWservice);
+                console.log("wServices", $scope.$parent.wServices);
+
+
+                //select entity
+                $scope.$parent.selectWservice(newWservice.id);
+                console.log("entities",$scope.$parent.entities);
             });
         }
       
       $scope.delete=function(){
+        id=$scope.$parent.selectedEntity.id;
+        for (i=0; i<$scope.entities.length;i++){
+            if ($scope.entities[i]){
+              if (id==$scope.entities[i].id){
 
-        var arrIndex=$scope.$parent.wServices.indexOf($scope.$parent.selectedWService);
-            $scope.$parent.wServices.splice(arrIndex,1);
 
-            WService.delete({
-              id:$scope.$parent.selectedWService.id
+                var arrIndex=$scope.$parent.wServices.indexOf($scope.entities[i]);
+                //delete from the entities array
+                $scope.$parent.entities.splice(i,1);
+                //delete from the cObjects array
+                $scope.$parent.wServices.splice(arrIndex,1);     
+                
+                WService.delete({
+                  id:id
 
-            });
-            console.log("deleting id "+ $scope.id);
-            $scope.$parent.selectCobject(null);
+                });
+                console.log("deleting id "+id);
+                
+                $state.go('edit');
+
+                break;
+              }
+            }
+          } 
         }
 
 
-        $scope.updateName=function(){
-            wServiceId=$scope.$parent.selectedWService.id; 
-            newname=$scope.$parent.selectedWService.name;
-            WService.update({id:wServiceId, name: newname});
+        $scope.updateName=function(id,name){
+            WService.update({id:id, name: name});
             //console.log("nameUpdated");
         }
-
     }
 ]);
 
@@ -70,7 +102,7 @@ wService.controller('wServiceCtrl', ['$scope','$route', '$routeParams','$locatio
         {
             id:'@id', 
             'salva': {method:'POST', params:{project:'@id'}, url:'/wService/create/' },
-            'find': {method:'GET', params:{id:'@id'}, url:'/wService/find/' },
+            'find': {method:'GET', params:{id:'@id'}, url:'/wService/wService/find/' },
             'getWSData': {method:'GET', params:{id:'@id'}, url:'/wService/getWSData/:id'},
             'getWSTrigger': {method:'GET', params:{id:'@id'}, url:'/wService/getWSTrigger/:id'},
             'getWSAction': {method:'GET', params:{id:'@id'}, url:'/wService/getWSAction/:id'},
